Rename to embed template plugin
