--- apps/builder/app/shared/copy-paste/plugins.ts
@@ -1,3 +1,0 @@
-export * as instance from "./plugin-instance";
-export * as template from "./plugin-template";
-export * as mardkown from "./plugin-markdown";
