+++ apps/builder/app/shared/copy-paste/use-copy-paste.ts
@@ -1,11 +1,11 @@
 import { useEffect } from "react";
 import { initCopyPaste } from "./init-copy-paste";
 import * as instance from "./plugin-instance";
-import * as template from "./plugin-template";
+import * as embedTemplate from "./plugin-embed-template";
 import * as markdown from "./plugin-markdown";
 
 export const useCopyPaste = () => {
   useEffect(() => {
-    return initCopyPaste([instance, template, markdown]);
+    return initCopyPaste([instance, embedTemplate, markdown]);
   }, []);
 };
